+++ R/fstrcapture.R
@@ -1,8 +1,11 @@
 #' Capture string tokens into a data frame
 #'
 # -------------------------------------------------------------------------
-#' `fstrcapture()` is a more efficient alternative for [strcapture()] when
-#' using Perl-compatible regular expressions
+#' [fstrcapture()] is a more efficient alternative for [strcapture()] when
+#' using Perl-compatible regular expressions. It is underpinned by the
+#' [regexpr()] function. Whilst [fstrcapture()] only returns the first
+#' occurrence of the captures in a string, [gstrcapture()], built upon
+#' [gregexpr()], will return all.
 #'
 # -------------------------------------------------------------------------
 #' @inheritParams utils::strcapture
@@ -14,7 +17,9 @@
 #' containing a column for each capture expression. The column types are
 #' inherited from proto, as are the names unless the captures themselves are
 #' named (in which case these are prioritised). Cases in x that do not match
-#' the pattern have NA in every column.
+#' the pattern have NA in every column. For [gstrcapture()] there is an
+#' additional column, `string_id`, which links the output to the relevant
+#' element of the input vector.
 #'
 # -------------------------------------------------------------------------
 #' @examples
@@ -26,6 +31,7 @@
 #' pattern <- "(?<first>[[:upper:]][[:lower:]]+) (?<last>[[:upper:]][[:lower:]]+)"
 #' proto <- data.frame(a="", b="")
 #' fstrcapture(notables, pattern, proto)
+#' gstrcapture(notables, pattern, proto)
 #'
 #' # from strcapture example ----------------------------------------------
 #' # if unnamed capture then proto names used
@@ -54,6 +60,11 @@ fstrcapture <- function(x, pattern, proto) {
         proto <- as.data.frame(replicate(length(names), "", simplify = FALSE))
 
     nomatch <- is.na(m) | m == -1L
+
+    start <- attr(m, "capture.start")
+    if (ncol(start) != length(proto))
+        stop("The number of captures in 'pattern' != 'length(proto)'.")
+
     if (all(nomatch)) {
         out <- matrix(NA_character_, length(m), length(proto))
     } else {
@@ -63,13 +74,70 @@ fstrcapture <- function(x, pattern, proto) {
         end[nomatch, ] <- start[nomatch, ] <- NA
         res <- substring(x, start, end)
         out <- matrix(res, length(m))
-        if (ncol(out) != length(proto))
-            stop("The number of captures in 'pattern' != 'length(proto)'.")
     }
     out <- conformToProto(out, proto)
     if (all(names == ""))
         return(out)
     names(out) <- names
+    out
+}
+
+#' @rdname fstrcapture
+#' @export
+gstrcapture <- function(x, pattern, proto) {
+
+    m <- gregexpr(pattern = pattern, text = x, perl = TRUE, useBytes = FALSE)
+    names <- attr(m[[1L]], "capture.names")
+    if (is.null(names))
+        stop("No captures within pattern.")
+
+    if (missing(proto))
+        proto <- as.data.frame(replicate(length(names), "", simplify = FALSE))
+
+    if (length(m) == 1L) {
+        m <- m[[1L]]
+        start <- attr(m, "capture.start")
+        if (ncol(start) != length(proto))
+            stop("The number of captures in 'pattern' != 'length(proto)'.")
+        nomatch <- is.na(m) | m == -1L
+        if (all(nomatch)) {
+            out <- matrix(NA_character_, length(m), length(proto))
+        } else {
+            length <- attr(m, "capture.length")
+            end <- start + length - 1L
+            end[nomatch, ] <- start[nomatch, ] <- NA
+            res <- substring(x, start, end)
+            out <- matrix(res, length(m))
+        }
+        out <- conformToProto(out, proto)
+        if (!all(names == ""))
+            names(out) <- names
+    } else {
+        id <- rep.int(seq_along(m), lengths(m))
+        out <- lapply(seq_along(m), function(i) {
+            mm <- m[[i]]
+            start <- attr(mm, "capture.start")
+            if (ncol(start) != length(proto))
+                stop("The number of captures in 'pattern' != 'length(proto)'.")
+            nomatch <- is.na(mm) | mm == -1L
+            if (all(nomatch)) {
+                out <- matrix(NA_character_, length(mm), length(proto))
+            } else {
+                length <- attr(mm, "capture.length")
+                end <- start + length - 1L
+                end[nomatch, ] <- start[nomatch, ] <- NA
+                res <- substring(x[i], start, end)
+                out <- matrix(res, length(mm))
+            }
+            out
+        })
+        out <- do.call(rbind, out)
+        out <- conformToProto(out, proto)
+        if (!all(names == ""))
+            names(out) <- names
+        out$string_id <- id
+    }
+
     out
 }