+++ NEWS.md
@@ -1,5 +1,7 @@
 # ympes (development version)
 
+- New function `gstrcapture()`.
+
 # ympes 1.7.0
 
 - `cc()` can once again send generative code for its output to the system